created the GamesEndpoints.cs file and added endpoints there
